Allow ceph_pool to set pg_num when auto_scaler set to warn
ceph_pool will only set the pg_num and pgp_num options when the auto_scaler is set to off.
While it makes sense not to set them when the auto_scaler is on, the auto_scaler can also be set to "warn" mode where it will not adjust the PG count either.
It seems counterintuitive to not allow configuring pg count in that case.

Signed-off-by: stryan <steve.ryan@joyent.com>